refactor owner login. move owner user into package.
