+++ core/domain/src/main/java/org/the_chance/honeymart/domain/usecase/usecaseManager/user/UserAuthenticationManagerUseCase.kt
@@ -1,4 +1,4 @@
-package org.the_chance.honeymart.domain.usecase.user
+package org.the_chance.honeymart.domain.usecase.usecaseManager.user
 
 import org.the_chance.honeymart.domain.usecase.LoginUserUseCase
 import org.the_chance.honeymart.domain.usecase.LogoutUserUseCase
