+++ core/domain/src/main/java/org/the_chance/honeymart/domain/usecase/usecaseManager/user/UserCouponsManagerUseCase.kt
@@ -1,4 +1,4 @@
-package org.the_chance.honeymart.domain.usecase.user
+package org.the_chance.honeymart.domain.usecase.usecaseManager.user
 
 import org.the_chance.honeymart.domain.usecase.ClipCouponUseCase
 import org.the_chance.honeymart.domain.usecase.GetClippedUserCouponsUseCase
