+++ user/src/main/java/org/the_chance/honeymart/ui/feature/cart/CartViewModel.kt
@@ -4,7 +4,7 @@ package org.the_chance.honeymart.ui.feature.cart
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.update
 import org.the_chance.honeymart.domain.model.Cart
-import org.the_chance.honeymart.domain.usecase.user.CartProductsManagerUseCase
+import org.the_chance.honeymart.domain.usecase.usecaseManager.user.CartProductsManagerUseCase
 import org.the_chance.honeymart.domain.util.ErrorHandler
 import org.the_chance.honeymart.ui.base.BaseViewModel
 import javax.inject.Inject
